Python/PDF-image-to-text/requirements.txt
pillow
pdf2image
pytesseract
glob